abandon "MVC" organizational pattern
We need to start using the "literate" in "literate programming" (LP)
more.  The problem with this "MVC" delineation is that it places a rigid
separation between how we categorize different components of the system.

Combine the MVC bits into a single "internals.org" file. We can split
and reognize this later as needed, focusing on LP and not MVC
architecture.