teach audit.sh to ignore certain files
Some files in this repo are tracked, but not tangled. Examples:

- All Org files fall into this category (as they are the source from
  which we tangle).

- All HTML files are woven from the Org code, not tangled, so they are
  also excluded.

The two categories above can be easily excluded from the audit. However
there are some other categories which must be hand-tuned. These
categories include:

- Generated files, e.g., gRPC stubs.

- LICENSE files (these are copied verbatim from the copyleft
  organizations that author them).

- Symlinks (such as to dependencies, like Lilac).

While doing this cleanup we've discovered some files which are either
deprecated (such as the files related to the "ptu" program which was
removed in b5d7b5e (remove ptu utility, 2024-07-06) or unused ("melbyd"
wrapper script which has fallen out of use). Delete these files.