update architecture diagram
The previous SVG was just a stub for checking the image generation
pipeline. Draw a more correct architecture diagram.

Note that we don't use dvisvgm yet, because we haven't tested it. But in
theory it should work, and when it does we probably won't need to use
pdf2svg and inkscape in the pipeline.

For now, we generate the images separately with `make build-images`,
separately from the rest of the weaving process.