images: control entire image generation pipeline
Instead of asking Org to _execute_ a "latex" block automatically to
generate a PDF with some magic, do it ourselves. This makes the code far
more easier to understand because we have a complete view of the exact
generated (tangled) "tex" file along with the commands to compile the
.tex to .pdf.

We use the LaTeX code that Org uses to generate standalone PDF files,
but invoke lualatex instead of pdflatex (Org default).

The Makefile isn't actually written very well, because we always
regenerate images even if the image.org file did not change. Add a FIXME
note for this.