simplify image generation stack
The dvisvgm bug [1] has been fixed. We can now just use dvisvgm directly
instead of having to use a combination of pdf2svg and inkscape. To be
fair, the use of inkscape is actually obviated by the "standalone" CTAN
package which removes the need to crop the PDF, which we learned while
trying to use dvisvgm directly.

IOW we could have continued to just use pdf2svg. But, let's use dvisvgm
because it's newer and also more closely tied to the LaTeX community.
Also pdf2svg appears to be unmaintained.

Remove unnecessary packages, both from the LaTeX code and also from
shell.nix.

[1] https://github.com/mgieseki/dvisvgm/issues/183